offer RESTful access to air quality data
Implemented a REST controller ofering the data in a convenint form, given
just the location for which the air quality data is requested. This is
much simpler, on the client's side, than using the original API directly,
as well as takes significantly lower amount of requests to get the same results.

The returned air quality index is built based on data from measurement
stations within approx. 10 km around the requested location. SubIndexes
missing on the nearest stations are taken from the farther ones. Details
are included in a form of an encoded PollutionDetails object.