fix DOS amplification issue
Removed a rest mapping allowing malicious users to put additional load
on the original data source used by the application. This was possible
via requesting a forced update of the data held by the application.
A single request from a malicious user would have resulted in over a
hundred requests sent to the original data source by the application.