replace @RestController with @Controller
Until now, /info mapping lead to a RestController which returned
dynamically generated html. This was replaced with a Controller, for
now the method returning the html is also annotated as @ResponseBody,
which together is equivalent to a single @RestController annotation.

Later on template can be used instead of generating the html in
the InfoService.